Yarn lockfile updated,
Apparently on Windows 7 using `npm install` does not seem
to succeed due to SASS not being able to find Visual Studio.

Unfortunately I was not able to find any good solution, but apparently
yarn was able to handle the problem much better, so I'm going to have to
use yarn to run and build the project.