Preprocessor permissive but working,
Will do a little bit of clean-up. Uses a "parsable" object which
allows doing lexical analysis easier so it can be reused for future
parsers.

While this preprocessor is very permissive, it doesn't seem to error out
on at least obvious error cases. In those cases it'll refuse to output
anything.