Naive MongoDB Atlas&Realm integration
Our articles are now fetched from the database. However this is the
most naive way I could implement it.

All other functionality works though.

Will continue to work on during the week, however it
suffices. This is really just a `marker` commit to use as a restore
point. The application now has a "load" buffer as there is a noticable
delay while logging into the DB (about a second...)