Patched for extraneous markup variations on Ubuntu 20.04 Firefox
I have also finally realized, that markdown-it, does not support
rendering underlines... I only now just realized this...

Anyways I don't know if I said this last commit, but I fixed the
synchronization issue with an extrahook to insertArticles.

Am I happy with it? Probably not. Was it simple? Simple enough, and at
least it's a small change so it's pretty easy to maintain.