RichTextEditor now maintains an initial state for it's text.
This is because I was annoyed at the contentEditable losing focus. It
seems to prevent the immediate article shimming, however it does appear
to be satisfactory enough. Until we have multiple people editting the
same article.

Of course, that will be a serious problem. I could always just specify a
lock system on the articles to prevent people from editting it... Just
hope that no one crashes I guess... I'll look about that later.