Working to add real authentication.
It's still not iron, but it's more secure than the previous method,
however I need to break and set some bones in the codebase to
accommodate this. Not too many, probably. I'm just going to do the
bare minimum to still have the feature functional.

New registering and login page for future things.

You have to pull from this branch in order to still have the website
working as intended.