getAllTags, and also faster load time by not awaiting so much.
I can allow the dispatches in initApi after loginWithPassword to run
asynchronously. Redux seems to make this relatively safe to do, so
this'll appear basically instant after the initial login. So this is
much faster appearing.

getAllTags was written into the MongoDB realm portion, it's basically
just a specialized query with some filtering and list flattening to get
all the tags the application is potentially aware of.