Hmmm, redux selector is not updating properly due to an identity issue?
Also due to lazy loading the websites (and therefore the SCSS), this
causes weird issues with the styling changing after going to a page on
the site.

We have to preload the SCSS to prevent this... Also we shouldn't really
be preloading Register and Login.

Admin should probably be the only site we lazy load for obvious reasons.