intentionally broken for now
Alright, so the RichTextEditor is starting to prove a little unwieldy,
and it's time to break the codebase.

I'm going to passively work on the Image editor to get it to have the
features I want without having to focus on text generation yet.

So right now, articles will not generate from this branch until later
because doing both at the same time... Is a bit painful...

So I'll rewrite the DOM to markdown at the end of these new features
because I now understand the limits of the features we require support
for...

It might be safe to re-enable the markdown editor for now to allow for
editting articles, so I'm probably going to be bringing that back from
an earlier branch, to make things little easier...

Basically this is turning into a fledgeling Media Wiki.