Context menu for both items working
While we cannot currently drag images with captions, that might be
acceptable, and also I'm saying that cause it'll be a real pain in the
ass to do it.

There's very weak table support for the Rich Text Editor and list
support unfortuantely, however most articles may not need it... Maybe if
someone complains we'll work through it...

I'm heavily leaning towards divising a fresh markup format not too
unlike MediaWiki, since normal markdown is too general to work for this.