Minor refactor of preprocessMarkdown
It's mostly shifting stuff out, since the code is likely
going to remain in behavior the same way.

I don't see a need to rewrite it that dramatically anymore, as it's
quite robust in handling it's cases.

While it could be shorter... It should be easy for me to jump in and
modify...