Strapi Integration testing
With a quick hour I only managed to setup a barebones Strapi quickstart
for Recyclopedia.

It's very very barebones, however the article index and article are read
from Strapi this time around instead of MongoDB, all other social
features... Are not working as intended!

This should still work with the static site generation because it
basically just wraps around what we already have!

I will continue work on this throughout the week as I feel like, but I
should be able to port things over by then.