Requests wrapper, for future stuff!
I wrote a wrapper around a HTTP request API. It's currently designed to
map to Axios because it's the one we currently use, but it's simple
enough to adapt for future things.

Using yarn workspace this should remove the amount of code we have to
write at least regarding queries and uploads.