Incremental Static Site Generation (#1)
Seems to work. Also it appears for dynamic routes I have literally no
indication of whether it actually works or not in the build, so I
probably spent 30 minutes trying to wonder if it worked because the
result from the build command doesn't tell me it worked...

I have no idea DOOD, all I know is that it works now.