More RTFM-ing, more in description.
Recycled Articles are interposed with normal Articles to be easier,
however the routes are expected to stay the same.

This may all be dumped for some GraphQL thingie, but this is more
explicit so this makes more sense to me.

Time to hook up all displays again.

Then I have to fix all the "non-handled" request errors.